Add 'filter' definition
