+++ requirements_test.txt
@@ -1,5 +1,5 @@
-mock>=1.3.0
-pytest~=3.8.2
+mock~=2.0
+pytest~=4.6
 pytest-cov~=2.4
 flake8~=2.5.1
 pep257~=0.7.0
