+++ tests/test_utils.py
@@ -82,7 +82,7 @@ class TestUtils(unittest.TestCase):
             parsed_node = parse_expression(condition_text)
             print(condition_node)
             print(parsed_node)
-            self.assertEquals(condition_node.render(), parsed_node.render(), *args)
+            self.assertEqual(condition_node.render(), parsed_node.render(), *args)
 
         assert_kv_match({"name": "net.exe"},
                         r"name == 'net.exe'",
@@ -146,25 +146,25 @@ class TestUtils(unittest.TestCase):
     def test_output_types(self):
         """Test that output types are correctly returned from eql.utils.get_output_types."""
         query_ast = parse_query("process where true")
-        self.assertEquals(get_output_types(query_ast), ["process"])
+        self.assertEqual(get_output_types(query_ast), ["process"])
 
         query_ast = parse_analytic({"query": "process where descendant of [file where true]"})
-        self.assertEquals(get_output_types(query_ast), ["process"])
+        self.assertEqual(get_output_types(query_ast), ["process"])
 
         query_ast = parse_query("file where true | unique pid | head 1")
-        self.assertEquals(get_output_types(query_ast), ["file"])
+        self.assertEqual(get_output_types(query_ast), ["file"])
 
         query_ast = parse_query("file where true | unique_count file_path")
-        self.assertEquals(get_output_types(query_ast), ["file"])
+        self.assertEqual(get_output_types(query_ast), ["file"])
 
         query_ast = parse_query("any where true | unique_count file_path")
-        self.assertEquals(get_output_types(query_ast), ["any"])
+        self.assertEqual(get_output_types(query_ast), ["any"])
 
         query_ast = parse_query("file where true | count")
-        self.assertEquals(get_output_types(query_ast), ["generic"])
+        self.assertEqual(get_output_types(query_ast), ["generic"])
 
         query_ast = parse_query("file where true | count process_name")
-        self.assertEquals(get_output_types(query_ast), ["generic"])
+        self.assertEqual(get_output_types(query_ast), ["generic"])
 
         query_ast = parse_query("""
         sequence
@@ -175,7 +175,7 @@ class TestUtils(unittest.TestCase):
             [process where true]
             [network where true]
         """)
-        self.assertEquals(get_output_types(query_ast), ["registry", "file", "process", "process", "process", "network"])
+        self.assertEqual(get_output_types(query_ast), ["registry", "file", "process", "process", "process", "network"])
 
         query_ast = parse_query("""
         sequence
@@ -188,7 +188,7 @@ class TestUtils(unittest.TestCase):
         | count event_type
         | head 5
         """)
-        self.assertEquals(get_output_types(query_ast), ["generic"])
+        self.assertEqual(get_output_types(query_ast), ["generic"])
 
         query_ast = parse_query("""
         sequence
@@ -203,4 +203,4 @@ class TestUtils(unittest.TestCase):
         | head 5
         | filter events[4].process_name == 'test.exe'
         """)
-        self.assertEquals(get_output_types(query_ast), ["registry", "file", "process", "process", "process", "network"])
+        self.assertEqual(get_output_types(query_ast), ["registry", "file", "process", "process", "process", "network"])
